refactor: extract backup coordinator and events
